# docs/API/Admin/Admin_swagger.yaml
basePath: /api/v1
definitions:
  admin.updateStatusRequest:
    properties:
      isEnabled:
        type: boolean
    type: object
...
  model.FlowMount:
    properties:
      createdAt:
        type: string
      eventKey:
        type: string
      flowID:
        type: integer
      isEnabled:
        type: boolean
      mountID:
        type: integer
      scope:
        type: string
      targetID:
        type: integer
      targetType:
        type: string
    type: object
  model.FlowStep:
    properties:
      failureStrategy:
        type: string
      flowID:
        type: integer
      inputTemplate:
        additionalProperties: {}
        type: object
      isEnabled:
        type: boolean
      scriptID:
        type: integer
      scriptVersionID:
        type: integer
      stepID:
        type: integer
      stepName:
        type: string
      stepOrder:
        type: integer
      timeoutMs:
        type: integer
    type: object
...
  model.Response:
    properties:
      code:
        description: 状态码，和http状态码剥离，以分离网络和业务
        example: 200
        type: integer
      msg:
        description: 信息。可以是NULL，也可以是成功时返回的结构体，也可以是失败时的错误信息
    type: object
  model.ScriptDefinition:
    properties:
      createdAt:
        type: string
      description:
        type: string
      interpreter:
        type: string
      isEnabled:
        type: boolean
      meta:
        additionalProperties: {}
        type: object
      scriptID:
        type: integer
      scriptKey:
        type: string
      scriptName:
        type: string
      updatedAt:
        type: string
    type: object
  model.ScriptFlow:
    properties:
      createdAt:
        type: string
      description:
        type: string
      flowID:
        type: integer
      flowKey:
        type: string
      flowName:
        type: string
      isEnabled:
        type: boolean
      meta:
        additionalProperties: {}
        type: object
      updatedAt:
        type: string
    type: object
  model.ScriptVersion:
    properties:
      checksum:
        type: string
      createdAt:
        type: string
      createdBy:
        type: integer
      fileName:
        type: string
      isActive:
        type: boolean
      relativePath:
        type: string
      scriptID:
        type: integer
      versionID:
        type: integer
      versionNum:
        type: integer
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Ubik 系统管理后台服务接口文档
  title: 管理后台 API
  version: "1.0"
paths:
...
  /admin/script-flows:
    get:
      consumes:
      - application/json
      description: 管理员获取脚本流程列表
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                msg:
                  items:
                    $ref: '#/definitions/model.ScriptFlow'
                  type: array
              type: object
      summary: 获取脚本流程列表
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: 管理员创建脚本流程
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 脚本流程
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.ScriptFlow'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                msg:
                  $ref: '#/definitions/model.ScriptFlow'
              type: object
      summary: 创建脚本流程
      tags:
      - Admin
  /admin/script-flows/{flow_id}:
    get:
      consumes:
      - application/json
      description: 管理员根据流程ID获取脚本流程
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 流程ID
        in: path
        name: flow_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                msg:
                  $ref: '#/definitions/model.ScriptFlow'
              type: object
      summary: 获取脚本流程详情
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: 管理员更新指定脚本流程
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 流程ID
        in: path
        name: flow_id
        required: true
        type: integer
      - description: 更新后的流程
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.ScriptFlow'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                msg:
                  $ref: '#/definitions/model.ScriptFlow'
              type: object
      summary: 更新脚本流程
      tags:
      - Admin
  /admin/script-flows/{flow_id}/mounts:
    get:
      consumes:
      - application/json
      description: 管理员获取指定流程的挂载列表
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 流程ID
        in: path
        name: flow_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                msg:
                  items:
                    $ref: '#/definitions/model.FlowMount'
                  type: array
              type: object
      summary: 获取流程挂载列表
      tags:
      - Admin
  /admin/script-flows/{flow_id}/status:
    post:
      consumes:
      - application/json
      description: 管理员更新指定流程的启用状态
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 流程ID
        in: path
        name: flow_id
        required: true
        type: integer
      - description: 状态请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/admin.updateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/model.Response'
      summary: 更新流程启用状态
      tags:
      - Admin
  /admin/script-flows/{flow_id}/steps:
    get:
      consumes:
      - application/json
      description: 管理员获取指定流程步骤
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 流程ID
        in: path
        name: flow_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                msg:
                  items:
                    $ref: '#/definitions/model.FlowStep'
                  type: array
              type: object
      summary: 获取流程步骤列表
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: 管理员替换指定流程的全部步骤
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 流程ID
        in: path
        name: flow_id
        required: true
        type: integer
      - description: 步骤列表
        in: body
        name: req
        required: true
        schema:
          items:
            $ref: '#/definitions/model.FlowStep'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/model.Response'
      summary: 替换流程步骤
      tags:
      - Admin
  /admin/script-flows/mounts:
    post:
      consumes:
      - application/json
      description: 管理员创建流程挂载
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 挂载信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.FlowMount'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                msg:
                  $ref: '#/definitions/model.FlowMount'
              type: object
      summary: 创建流程挂载
      tags:
      - Admin
  /admin/script-flows/mounts/{mount_id}:
    delete:
      consumes:
      - application/json
      description: 管理员删除指定流程挂载
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 挂载ID
        in: path
        name: mount_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/model.Response'
      summary: 删除流程挂载
      tags:
      - Admin
  /admin/scripts:
    get:
      consumes:
      - application/json
      description: 管理员获取脚本定义列表
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                msg:
                  items:
                    $ref: '#/definitions/model.ScriptDefinition'
                  type: array
              type: object
      summary: 获取脚本定义列表
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: 管理员创建脚本定义
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 脚本定义
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.ScriptDefinition'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                msg:
                  $ref: '#/definitions/model.ScriptDefinition'
              type: object
      summary: 创建脚本定义
      tags:
      - Admin
  /admin/scripts/{script_id}:
    get:
      consumes:
      - application/json
      description: 管理员根据脚本ID获取脚本定义
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 脚本ID
        in: path
        name: script_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                msg:
                  $ref: '#/definitions/model.ScriptDefinition'
              type: object
      summary: 获取脚本定义详情
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: 管理员更新指定脚本定义
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 脚本ID
        in: path
        name: script_id
        required: true
        type: integer
      - description: 更新后的脚本定义
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.ScriptDefinition'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                msg:
                  $ref: '#/definitions/model.ScriptDefinition'
              type: object
      summary: 更新脚本定义
      tags:
      - Admin
  /admin/scripts/{script_id}/status:
    post:
      consumes:
      - application/json
      description: 管理员更新指定脚本的启用状态
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 脚本ID
        in: path
        name: script_id
        required: true
        type: integer
      - description: 状态请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/admin.updateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/model.Response'
      summary: 更新脚本启用状态
      tags:
      - Admin
  /admin/scripts/{script_id}/versions:
    get:
      consumes:
      - application/json
      description: 管理员获取指定脚本的版本列表
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 脚本ID
        in: path
        name: script_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                msg:
                  items:
                    $ref: '#/definitions/model.ScriptVersion'
                  type: array
              type: object
      summary: 获取脚本版本列表
      tags:
      - Admin
  /admin/scripts/{script_id}/versions/{version_id}/activate:
    post:
      consumes:
      - application/json
      description: 管理员激活指定脚本版本
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 脚本ID
        in: path
        name: script_id
        required: true
        type: integer
      - description: 版本ID
        in: path
        name: version_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 激活成功
          schema:
            $ref: '#/definitions/model.Response'
      summary: 激活脚本版本
      tags:
      - Admin
  /admin/scripts/{script_id}/versions/upload:
    post:
      consumes:
      - multipart/form-data
      description: 管理员上传脚本文件并创建新版本
      parameters:
      - description: Bearer {access_token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: 脚本ID
        in: path
        name: script_id
        required: true
        type: integer
      - description: 脚本文件
        in: formData
        name: script_file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/model.Response'
            - properties:
                msg:
                  $ref: '#/definitions/model.ScriptVersion'
              type: object
      summary: 上传脚本版本
      tags:
      - Admin
...
swagger: "2.0"